Moved match-making logic to utility file
